pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "streamt"
version = "0.1.0"
description = "dbt for streaming - declarative streaming pipelines"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "streamt contributors" }
]
keywords = ["kafka", "streaming", "dbt", "flink", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "confluent-kafka>=2.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "types-pyyaml",
    "types-requests",
]

[project.scripts]
streamt = "streamt.cli:main"

[project.urls]
Homepage = "https://github.com/conduktor/streamt"
Documentation = "https://github.com/conduktor/streamt"
Repository = "https://github.com/conduktor/streamt"

[tool.hatch.build.targets.wheel]
packages = ["src/streamt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
